Move atom_desktop_window_tree_host_win to atom/browser/ui/win directory.
