+++ atom/browser/ui/win/atom_desktop_window_tree_host_win.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#include "atom/browser/atom_desktop_window_tree_host_win.h"
+#include "atom/browser/ui/win/atom_desktop_window_tree_host_win.h"
 #include "atom/browser/ui/win/thumbar_host.h"
 
 namespace atom {
