+++ atom/browser/ui/win/atom_desktop_window_tree_host_win.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#ifndef ATOM_BROWSER_ATOM_DESKTOP_WINDOW_TREE_HOST_WIN_H_
-#define ATOM_BROWSER_ATOM_DESKTOP_WINDOW_TREE_HOST_WIN_H_
+#ifndef ATOM_BROWSER_UI_WIN_ATOM_DESKTOP_WINDOW_TREE_HOST_WIN_H_
+#define ATOM_BROWSER_UI_WIN_ATOM_DESKTOP_WINDOW_TREE_HOST_WIN_H_
 
 #include <windows.h>
 
@@ -38,4 +38,4 @@ class AtomDesktopWindowTreeHostWin : public views::DesktopWindowTreeHostWin {
 
 }  // namespace atom
 
-#endif  // ATOM_BROWSER_ATOM_DESKTOP_WINDOW_TREE_HOST_WIN_H_
+#endif  // ATOM_BROWSER_UI_WIN_ATOM_DESKTOP_WINDOW_TREE_HOST_WIN_H_
